refactor: Keystore and key module
TODO: update and fix tests.
