+++ packages/neuron-wallet/src/keys/key.ts
@@ -1,48 +1,58 @@
 import crypto from 'crypto'
-import SHA3 from 'sha3'
-import { v4 as uuid } from 'uuid'
-import { ec as EC } from 'elliptic'
 
 import Address, { HDAddress } from '../services/addresses'
-import { Keystore, KdfParams } from './keystore'
-import { Keychain } from './hd'
+import Keystore from './keystore'
+import Keychain, { privateToPublic } from './keychain'
 import { Validate, Required, Password } from '../decorators'
-import { IncorrectPassword, IsRequired, InvalidMnemonic, UnsupportedCipher } from '../exceptions'
-import { entropyToMnemonic, validateMnemonic, mnemonicToSeed } from '../utils/mnemonic'
-
-const ec = new EC('secp256k1')
+import { IsRequired, InvalidMnemonic } from '../exceptions'
+import { entropyToMnemonic, validateMnemonic, mnemonicToSeed } from './mnemonic'
 
 const ENTROPY_SIZE = 16
 
-// Private/Public key and chain code.
-export class ExtendedKey {
-  privateKey?: string
-  publicKey?: string
+export class ExtendedPublicKey {
+  publicKey: string
   chainCode: string
 
-  constructor(privateKey: string | undefined, publicKey: string | undefined, chainCode: string) {
-    this.privateKey = privateKey
+  constructor(publicKey: string, chainCode: string) {
     this.publicKey = publicKey
     this.chainCode = chainCode
   }
 
-  serializePrivate = () => {
-    return this.privateKey! + this.chainCode
+  serialize = () => {
+    return this.publicKey + this.chainCode
   }
 
-  serializePublic = () => {
-    return this.privateKey! + this.chainCode
+  static parse = (serialized: string) => {
+    return new ExtendedPublicKey(serialized.slice(0, 66), serialized.slice(66))
+  }
+}
+
+export class ExtendedPrivateKey {
+  privateKey: string
+  chainCode: string
+
+  constructor(privateKey: string, chainCode: string) {
+    this.privateKey = privateKey
+    this.chainCode = chainCode
   }
 
-  static parsePrivate = (serialized: string) => {
-    const privateKey = serialized.slice(0, 64)
-    const publicKey = ec.keyFromPrivate(privateKey).getPublic(true, 'hex') as string
-    return new ExtendedKey(privateKey, publicKey, serialized.slice(64))
+  serialize = () => {
+    return this.privateKey + this.chainCode
   }
 
-  static parsePublic = (serialized: string) => {
-    return new ExtendedKey(undefined, serialized.slice(0, 66), serialized.slice(66))
+  toExtendedPublicKey = (): ExtendedPublicKey => {
+    const publicKey = privateToPublic(Buffer.from(this.privateKey, 'hex')).toString('hex')
+    return new ExtendedPublicKey(publicKey, this.chainCode)
   }
+
+  static parse = (serialized: string) => {
+    return new ExtendedPrivateKey(serialized.slice(0, 64), serialized.slice(64))
+  }
+}
+
+export enum DefaultAddressNumber {
+  Receiving = 20,
+  Change = 10,
 }
 
 export interface Addresses {
@@ -50,32 +60,12 @@ export interface Addresses {
   change: HDAddress[]
 }
 
-enum DefaultAddressNumber {
-  Receiving = 20,
-  Change = 10,
-}
-
 export default class Key {
-  public mnemonic?: string
   public keystore?: Keystore
-  public extendedKey?: ExtendedKey
   public addresses?: Addresses
 
-  constructor({
-    mnemonic,
-    keystore,
-    extendedKey,
-    addresses,
-  }: {
-    mnemonic?: string
-    keystore?: Keystore
-    extendedKey?: ExtendedKey
-    addresses?: Addresses
-  } = {}) {
-    this.mnemonic = mnemonic
+  constructor(keystore: Keystore | undefined) {
     this.keystore = keystore
-    this.extendedKey = extendedKey
-    this.addresses = addresses
   }
 
   static generateMnemonic = () => {
@@ -84,41 +74,10 @@ export default class Key {
   }
 
   @Validate
-  static async fromKeystore(@Required keystore: string, @Password password: string) {
-    const keystoreObject: Keystore = JSON.parse(keystore)
-    const key = new Key()
-    key.keystore = keystoreObject
-    if (!key.checkPassword(password)) {
-      throw new IncorrectPassword()
-    }
-    const { kdfparams } = keystoreObject.crypto
-    const derivedKey: Buffer = crypto.scryptSync(
-      Buffer.from(password),
-      Buffer.from(kdfparams.salt, 'hex'),
-      kdfparams.dklen,
-      {
-        N: kdfparams.n,
-        r: kdfparams.r,
-        p: kdfparams.p,
-      }
-    )
-    const ciphertext = Buffer.from(keystoreObject.crypto.ciphertext, 'hex')
-    const mac = new SHA3(256)
-      .update(Buffer.concat([derivedKey.slice(16, 32), ciphertext]))
-      .digest()
-      .toString('hex')
-    if (mac !== keystoreObject.crypto.mac) {
-      throw new Error('Key derivation failed - possibly wrong password')
-    }
-    const decipher = crypto.createDecipheriv(
-      keystoreObject.crypto.cipher,
-      derivedKey.slice(0, 16),
-      Buffer.from(keystoreObject.crypto.cipherparams.iv, 'hex')
-    )
-    const encrypted = Buffer.concat([decipher.update(ciphertext), decipher.final()]).toString('hex')
-    const extendedKey = ExtendedKey.parsePrivate(encrypted)
-    key.extendedKey = extendedKey
-    key.addresses = Address.generateAddresses(extendedKey, DefaultAddressNumber.Receiving, DefaultAddressNumber.Change)
+  static async fromKeystore(@Required keystore: Keystore, @Password password: string) {
+    const extendedPrivateKey = keystore.extendedPrivatKey(password)
+    const key = new Key(keystore)
+    key.addresses = Address.generateAddresses(extendedPrivateKey.toExtendedPublicKey())
     return key
   }
 
@@ -127,14 +86,16 @@ export default class Key {
     if (!validateMnemonic(mnemonic)) {
       throw new InvalidMnemonic()
     }
-    const key = new Key()
-    key.extendedKey = await key.extendedKeyFromMnemonic(mnemonic)
-    key.addresses = Address.generateAddresses(
-      key.extendedKey,
-      DefaultAddressNumber.Receiving,
-      DefaultAddressNumber.Change
-    )
-    key.keystore = key.toKeystore(key.extendedKey.serializePrivate(), password)
+
+    const seed = await mnemonicToSeed(mnemonic)
+    const root = Keychain.fromSeed(seed)
+    if (!root.privateKey) {
+      throw new InvalidMnemonic()
+    }
+    const extendedKey = new ExtendedPrivateKey(root.privateKey.toString('hex'), root.chainCode.toString('hex'))
+    const keystore = Keystore.create(extendedKey, password)
+    const key = new Key(keystore)
+    key.addresses = Address.generateAddresses(extendedKey.toExtendedPublicKey())
     return key
   }
 
@@ -145,95 +106,7 @@ export default class Key {
     if (this.keystore === undefined) {
       throw new IsRequired('Keystore')
     }
-    const { kdfparams } = this.keystore.crypto
-    const derivedKey: Buffer = crypto.scryptSync(
-      Buffer.from(password),
-      Buffer.from(kdfparams.salt, 'hex'),
-      kdfparams.dklen,
-      {
-        N: kdfparams.n,
-        r: kdfparams.r,
-        p: kdfparams.p,
-      }
-    )
-    const ciphertext = Buffer.from(this.keystore.crypto.ciphertext, 'hex')
-    const mac = new SHA3(256)
-      .update(Buffer.concat([derivedKey.slice(16, 32), ciphertext]))
-      .digest()
-      .toString('hex')
-    return mac === this.keystore.crypto.mac
-  }
 
-  public nextUnusedAddress = () => {
-    if (this.extendedKey) {
-      return Address.nextUnusedAddress(this.extendedKey)
-    }
-    return ''
-  }
-
-  public allUsedAddresses = () => {
-    if (this.extendedKey) {
-      return Address.searchUsedAddresses(this.extendedKey)
-    }
-    return []
-  }
-
-  private extendedKeyFromMnemonic = async (mnemonic: string) => {
-    const seed = await mnemonicToSeed(mnemonic)
-    const root = Keychain.fromSeed(seed)
-    if (root.privateKey) {
-      return new ExtendedKey(
-        root.privateKey.toString('hex'),
-        root.publicKey.toString('hex'),
-        root.chainCode.toString('hex')
-      )
-    }
-    throw new InvalidMnemonic()
-  }
-
-  public toKeystore = (encryptedData: string, password: string) => {
-    const salt = crypto.randomBytes(32)
-    const iv = crypto.randomBytes(16)
-    const params = {
-      n: 8192,
-      r: 8,
-      p: 1,
-    }
-    const kdfparams: KdfParams = {
-      dklen: 32,
-      salt: salt.toString('hex'),
-      ...params,
-    }
-    const derivedKey: Buffer = crypto.scryptSync(Buffer.from(password), salt, kdfparams.dklen, {
-      N: kdfparams.n,
-      r: kdfparams.r,
-      p: kdfparams.p,
-    })
-
-    const cipher = crypto.createCipheriv('aes-128-ctr', derivedKey.slice(0, 16), iv)
-    if (!cipher) {
-      throw new UnsupportedCipher()
-    }
-    const ciphertext = Buffer.concat([cipher.update(Buffer.from(encryptedData, 'utf8')), cipher.final()])
-    const hash = new SHA3(256)
-    const mac = hash
-      .update(Buffer.concat([derivedKey.slice(16, 32), ciphertext]))
-      .digest()
-      .toString('hex')
-
-    return {
-      version: 3,
-      id: uuid(),
-      crypto: {
-        ciphertext: ciphertext.toString('hex'),
-        cipherparams: {
-          iv: iv.toString('hex'),
-        },
-        cipher: 'aes-128-ctr',
-        kdf: 'scrypt',
-        kdfparams,
-        mac,
-      },
-    }
+    return this.keystore.checkPassword(password)
   }
 }
