+++ packages/neuron-wallet/src/keys/keystore.ts
@@ -1,3 +1,12 @@
+import crypto from 'crypto'
+import SHA3 from 'sha3'
+import { v4 as uuid } from 'uuid'
+
+import { UnsupportedCipher, IncorrectPassword } from '../exceptions'
+import { ExtendedPrivateKey } from './key'
+
+const CIPHER = 'aes-128-ctr'
+
 export interface CipherParams {
   iv: string
 }
@@ -19,8 +28,115 @@ export interface Crypto {
   mac: string
 }
 
-export interface Keystore {
+// Encrypt and save master extended private key.
+export default class Keystore {
   crypto: Crypto
   id: string
-  version: number
+  version: number = 3
+
+  constructor(theCrypto: Crypto, id: string) {
+    this.crypto = theCrypto
+    this.id = id
+  }
+
+  static create = (extendedPrivateKey: ExtendedPrivateKey, password: string) => {
+    const salt = crypto.randomBytes(32)
+    const iv = crypto.randomBytes(16)
+    const params = {
+      n: 8192,
+      r: 8,
+      p: 1,
+    }
+    const kdfparams: KdfParams = {
+      dklen: 32,
+      salt: salt.toString('hex'),
+      ...params,
+    }
+    const derivedKey: Buffer = crypto.scryptSync(Buffer.from(password), salt, kdfparams.dklen, {
+      N: kdfparams.n,
+      r: kdfparams.r,
+      p: kdfparams.p,
+    })
+
+    const cipher = crypto.createCipheriv(CIPHER, derivedKey.slice(0, 16), iv)
+    if (!cipher) {
+      throw new UnsupportedCipher()
+    }
+    const ciphertext = Buffer.concat([
+      cipher.update(Buffer.from(extendedPrivateKey.serialize(), 'utf8')),
+      cipher.final(),
+    ])
+    const hash = new SHA3(256)
+    const mac = hash
+      .update(Buffer.concat([derivedKey.slice(16, 32), ciphertext]))
+      .digest()
+      .toString('hex')
+
+    return new Keystore(
+      {
+        ciphertext: ciphertext.toString('hex'),
+        cipherparams: {
+          iv: iv.toString('hex'),
+        },
+        cipher: CIPHER,
+        kdf: 'scrypt',
+        kdfparams,
+        mac,
+      },
+      uuid()
+    )
+  }
+
+  // Decrypt and return serialized extended private key.
+  decrypt(password: string): string {
+    const { kdfparams } = this.crypto
+    const derivedKey: Buffer = crypto.scryptSync(
+      Buffer.from(password),
+      Buffer.from(kdfparams.salt, 'hex'),
+      kdfparams.dklen,
+      {
+        N: kdfparams.n,
+        r: kdfparams.r,
+        p: kdfparams.p,
+      }
+    )
+    const ciphertext = Buffer.from(this.crypto.ciphertext, 'hex')
+    const mac = new SHA3(256)
+      .update(Buffer.concat([derivedKey.slice(16, 32), ciphertext]))
+      .digest()
+      .toString('hex')
+    if (mac !== this.crypto.mac) {
+      throw new IncorrectPassword()
+    }
+    const decipher = crypto.createDecipheriv(
+      this.crypto.cipher,
+      derivedKey.slice(0, 16),
+      Buffer.from(this.crypto.cipherparams.iv, 'hex')
+    )
+    return Buffer.concat([decipher.update(ciphertext), decipher.final()]).toString('hex')
+  }
+
+  extendedPrivatKey = (password: string): ExtendedPrivateKey => {
+    return ExtendedPrivateKey.parse(this.decrypt(password))
+  }
+
+  checkPassword = (password: string) => {
+    const { kdfparams } = this.crypto
+    const derivedKey: Buffer = crypto.scryptSync(
+      Buffer.from(password),
+      Buffer.from(kdfparams.salt, 'hex'),
+      kdfparams.dklen,
+      {
+        N: kdfparams.n,
+        r: kdfparams.r,
+        p: kdfparams.p,
+      }
+    )
+    const ciphertext = Buffer.from(this.crypto.ciphertext, 'hex')
+    const mac = new SHA3(256)
+      .update(Buffer.concat([derivedKey.slice(16, 32), ciphertext]))
+      .digest()
+      .toString('hex')
+    return mac === this.crypto.mac
+  }
 }
