+++ packages/neuron-wallet/src/keys/mnemonic/word-list.ts
