+++ packages/neuron-wallet/tests/keys/mnemonic/fixtures.json
